l,h,n bonus and norm done, gonna start #, ,+

+++ ft_printf/get_uintarg.c
@@ -1,9 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   get_uintarg.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: soekim <soekim@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/22 15:37:54 by soekim            #+#    #+#             */
+/*   Updated: 2021/02/22 15:40:33 by soekim           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "ft_printf.h"
 
 unsigned long long	get_uintarg(va_list param, int memlen)
 {
-	int i;
+	unsigned long long		i;
 
 	i = 0;
 	if (memlen == 0)
